+++ src/main/java/org/ladocuploader/app/submission/actions/EncryptSSNBeforeSaving.java
@@ -3,7 +3,6 @@ package org.ladocuploader.app.submission.actions;
 import formflow.library.config.submission.Action;
 import formflow.library.data.Submission;
 import org.ladocuploader.app.submission.StringEncryptor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -11,9 +10,8 @@ public class EncryptSSNBeforeSaving implements Action {
 
   private final StringEncryptor encryptor;
 
-  public EncryptSSNBeforeSaving(@Value("${form-flow.aws.cmk}") String keyArn, @Value("${form-flow.aws.access_key}") String accessKey,
-                                    @Value("${form-flow.aws.secret_key}") String secretKey) {
-    encryptor = new StringEncryptor(keyArn, accessKey, secretKey);
+  public EncryptSSNBeforeSaving(StringEncryptor stringEncryptor) {
+    encryptor = stringEncryptor;
   }
 
   public void run(Submission submission) {
