+++ src/main/resources/application.yaml
@@ -3,7 +3,6 @@ form-flow:
     path: ~
   inputs: 'org.ladocuploader.app.inputs.'
   path: 'flows-config.yaml'
-  encryption_key: ${ENCRYPTION_KEYSET}
   uploads:
     accepted-file-types: '.jpeg,.jpg,.png,.pdf,.bmp,.gif,.doc,.docx,.odt,.ods,.odp'
     thumbnail-width: '64'
